# Speedup: Replace modules and use lazy imports