--- a/setup.cfg
+++ b/setup.cfg
@@ -35,6 +35,7 @@ packages = find:
 install_requires =
     httpx>=0.19
     platformdirs
+    prettytable>=2.4.0
     pytablewriter[html]>=0.63
     python-dateutil
     python-slugify