--- a/setup.cfg
+++ b/setup.cfg
@@ -39,6 +39,7 @@ install_requires =
     pytablewriter[html]>=0.63
     python-dateutil
     python-slugify
+    importlib-metadata;python_version < '3.8'
 python_requires = >=3.6
 package_dir = =src
 setup_requires =