Checking the import times (see https://medium.com/alan/how-we-improved-our-python-backend-start-up-time-2c33cd4873c8) with:

```sh
python -m pip install tuna

python -X importtime -c 'import pypistats; pypistats.recent("pillow")' 2> import0.log
tuna import0.log
```

This test case is when using default Markdown format and the data has been cached.

# `main`

![image](https://user-images.githubusercontent.com/1324225/140051563-7438a3e8-f2b6-4fa8-aea7-fa3663010a6b.png)

1.014s total: shows importing pytablewriter (0.433s, 42.7%), pkg_resources (0.270s, 26.7%) and httpx (0.243s, 24.0%) are slow.


# Use faster PrettyTable for Markdown

![image](https://user-images.githubusercontent.com/1324225/140051813-cdb21fbc-8f87-4d96-84fa-508a237b33d2.png)

Use PrettyTable for default Markdown, use pytablewriter for the more complex formats. Only import each when needed.

We remove 0.433s for pytablewriter but only add 0.011s for PrettyTable.

Down to 0.627s.

# Use faster `importlib.metadata` for getting version

![image](https://user-images.githubusercontent.com/1324225/140051886-fe474d82-957b-4bba-b4c6-927ab526c537.png)

Instead of importing `pkg_resources` to get the package version (because we use `setuptools_scm`), use the much faster `importlib.metadata` (and backport for pre-3.8).

https://github.com/pypa/setuptools_scm#retrieving-package-version-at-runtime

We remove 0.272s for `pkg_resources` but only add 0.010s for `importlib.metadata`.

Down to 0.378s.

# Only import httpx if needed

If we happen to be reading a cached data file, we don't need to import httpx to fetch a new one.

![image](https://user-images.githubusercontent.com/1324225/140057803-5e553e19-720f-4433-9bc8-aa301575b336.png)

Down to 0.139s.

1.014s -> 0.139s = 7.3x faster.

# Another test

```console
$ python --version
Python 3.10.0
```
```console
$ git checkout main
Already on 'main'
Your branch is up to date with 'origin/main'.
$ time python -m timeit -n 1000 'import pypistats; pypistats.recent("pillow")'
1000 loops, best of 5: 18.8 msec per loop
python3 -m timeit -n 1000 'import pypistats; pypistats.recent("pillow")'  91.74s user 1.75s system 94% cpu 1:38.99 total
```
```console
$ git checkout speedup
Switched to branch 'speedup'
Your branch is up to date with 'origin/speedup'.
$ time python -m timeit -n 1000 'import pypistats; pypistats.recent("pillow")'
1000 loops, best of 5: 426 usec per loop
python3 -m timeit -n 1000 'import pypistats; pypistats.recent("pillow")'  2.21s user 0.50s system 91% cpu 2.951 total
```

= 44x faster per loop.

# And another

```console
$ git checkout main
Already on 'main'
Your branch is up to date with 'origin/main'.
$ time pypistats recent pillow
| last_day  | last_month | last_week |
| --------: | ---------: | --------: |
| 1,362,501 | 35,373,628 | 8,292,701 |

pypistats recent pillow  1.53s user 0.35s system 101% cpu 1.846 total
```
```console
$ gco speedup
Switched to branch 'speedup'
Your branch is up to date with 'origin/speedup'.
$ time pypistats recent pillow
|  last_day | last_month | last_week |
|----------:|-----------:|----------:|
| 1,362,501 | 35,373,628 | 8,292,701 |

pypistats recent pillow  0.16s user 0.06s system 90% cpu 0.244 total
```

= 9.5625x faster.